% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpreview.R
\name{ggpreview}
\alias{ggpreview}
\title{Preview a file that would be created by \code{ggsave()}}
\usage{
ggpreview(plot = ggplot2::last_plot(), ..., device = "png")
}
\arguments{
\item{plot}{Plot to view (default is last displayed plot).}

\item{...}{options passed onto \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}

\item{device}{the file extension of the device to use. Defaults to \code{"png"}.}
}
\description{
This function saves a plot to a temporary file with \code{ggsave()} and opens the
temporary file in the system viewer. This function is useful for quickly
previewing how a plot will look when it is saved to a file. It is adapted
from TJ Mahr's package
\href{https://github.com/tjmahr/tjmisc/blob/master/R/ggplot.R}{'tjmisc'}.
}
\examples{

if (interactive()) {
  library(ggplot2)
  library(magrittr)

  pl <- ggplot(mtcars, aes(x = cyl, y = mpg)) +
    geom_point()

  pl \%>\%
    ggpreview(width = 1.619 * 6.6,
              height = 7,
              dpi = 300,
              bg = "white")
}

}
