% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_figures.R
\name{save_figure.flextable}
\alias{save_figure.flextable}
\title{Save Flextable Figures}
\usage{
\method{save_figure}{flextable}(.figure, .path, create_directory = TRUE, ...)
}
\arguments{
\item{.figure}{The figure to be saved.}

\item{.path}{The path to save the figure (including file extension such as
".png").}

\item{create_directory}{If the \code{.path} argument specifies a non-existent
directory, should it be created automatically? Default is \code{TRUE}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The file path for the saved object.
}
\description{
Save 'flextable' objects. This is a wrapper around flextable's
\verb{save_as_*()} functions with the appropriate one chosen based on \code{.path}
file extension.
}
\examples{

if (interactive()) {

  library(flextable)
  library(magrittr)

  # As PNG
  mtcars \%>\%
    flextable() \%>\%
    save_figure("~/Desktop/mtcars-table.png")

  # As HTML
  mtcars \%>\%
    flextable() \%>\%
    save_figure("~/Desktop/mtcars-table.html")

}


}
