% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_figures.R
\name{save_figure.ggplot}
\alias{save_figure.ggplot}
\title{Save ggplot2 Figures}
\usage{
\method{save_figure}{ggplot}(
  .figure,
  .path,
  create_directory = TRUE,
  width = 1.619 * height,
  height = 4.94,
  units = "in",
  ...
)
}
\arguments{
\item{.figure}{The figure to be saved.}

\item{.path}{The path to save the figure (including file extension such as
".png").}

\item{create_directory}{If the \code{.path} argument specifies a non-existent
directory, should it be created automatically? Default is \code{TRUE}.}

\item{width}{The width of the plot (default is 1.619 * height)}

\item{height}{The height of the plot (default is 4.94)}

\item{units}{Units to save the plot in. Default is "in" (inches). See
\link[ggplot2:ggsave]{ggplot2::ggsave} for more details.}

\item{...}{Additional parameters to pass into \link[ggplot2:ggsave]{ggplot2::ggsave}.}
}
\value{
The file path for the saved object.
}
\description{
Save ggplot2 Figures
}
\examples{

if (interactive()) {
  library(ggplot2)

  car_plot <- ggplot(mtcars, aes(x = gear, y = cyl)) +
    geom_col()

  save_figure(car_plot, "~/Desktop/carplot.png")
}

}
